Fix up tests even more.
